ecdsa: rename `signing` and `verifying` modules
Renames modules as follows:

- `sign` => `signing`
- `verify` => `verifying`

This better matches the `SigningKey` and `VerifyingKey` type names.
